Updating passport view so the QR code generator and Print buttons are only available when an authenticated user is viewing their own passport. Created new MyPassport.jsx page for authenticated lookup and ViewPassport.jsx is for public lookup. CommunicationPassport.jsx component was created to be used on both pages and anywhere else it may be needed in the future.
